GH-47819: [CI][Packaging][Release] Avoid triggering Linux packages on release branch push (#47826)
### Rationale for this change

We require the Linux package jobs to be triggered on RC tag creation. For example for 22.0.0, we currently push the tag `apache-arrow-22.0.0-rc0` and the release branch `release-22.0.0-rc0`. Those events are triggering builds over the same commit and the tag event gets cancelled due to a "high priority task" triggering the same jobs. This causes jobs to fail on the branch because the ARROW_VERSION is not generated. If we manually re-trigger the jobs on the tag they are successful.

### What changes are included in this PR?

Remove the `release-*` branches from triggering the event to allow only the tag to run the jobs so they don't get cancelled.

### Are these changes tested?

No

### Are there any user-facing changes?

No

* GitHub Issue: #47819

Authored-by: Raúl Cumplido <raulcumplido@gmail.com>
Signed-off-by: Raúl Cumplido <raulcumplido@gmail.com>