GH-47830: [Release] Run RC verification source testing step in a subshell (#47831)
### Rationale for this change

Isolates environment variable changes made when verifying the source from the tests that verify the binaries, to prevent these causing errors like in #47830.

### What changes are included in this PR?

Changes `verify-release-candidate.sh` to run `test_source_distribution` in a subshell.

### Are these changes tested?

I've tested this manually, I'm not sure if there are any existing CI tests that run this script on PRs.

### Are there any user-facing changes?

No
* GitHub Issue: #47830

Authored-by: Adam Reeve <adreeve@gmail.com>
Signed-off-by: Adam Reeve <adreeve@gmail.com>