MINOR: [Docs] Add note to Windows docs about patch util (#47839)
### Rationale for this change

For certain C++ build configurations, we require `patch` to be in the user's `PATH`. This was introduced in https://github.com/apache/arrow/commit/e1f727cbb447d2385949a54d8f4be2fdc6cefe29 and a [Discussion pointed this out](https://github.com/apache/arrow/discussions/47829#discussioncomment-14702132). For Linux and macOS, it's pretty reasonable for us to assume `patch` is available but not on Windows.

### What changes are included in this PR?

Adds a note in the "Debug" section of the Windows docs.

### Are these changes tested?

Yes, I'll add a screenshot to make review easier.

### Are there any user-facing changes?

No.

Authored-by: Bryce Mecum <petridish@gmail.com>
Signed-off-by: AlenkaF <frim.alenka@gmail.com>