GH-42173: [R][C++] Writing partitioned dataset on S3 fails if ListBucket is not allowed for the user (#47599)
### Rationale for this change

This PR gives the user to choose not to create directory in the bucket before writing dataset.
In case the `create_directory` option is set to FALSE, no verification will be made by R arrow.
The S3 storage will itself verify if the directory exists and if the users has the rigth to modify it.
This way no `ListBucket`  or ` HeadBucket`  are necessary to achieve the write operation. 

``` 
df |> arrow::write_dataset(
  minio$path(paste0("smartsla-bucket/rarrow/")),
  partitioning = "qualitative",
  create_directory = FALSE,
  format = "parquet"
)
```

### What changes are included in this PR?

`create_directory` is now available to the user in the `write_dataset` function.
Before this PR, this option was automatically set to TRUE (by default).

### Are these changes tested?

Yes

### Are there any user-facing changes?

No, the default value for `create_directory` is still TRUE.

* GitHub Issue: #42173

Lead-authored-by: Simon ELBAZ <elbazsimon9@gmail.com>
Co-authored-by: Simon Elbaz <elbazsimon9@gmail.com>
Co-authored-by: Antoine Pitrou <pitrou@free.fr>
Signed-off-by: Antoine Pitrou <antoine@python.org>