GH-46147: [C++] Implement GCS support in Meson (#47568)
### Rationale for this change

This continues adding features to the Meson configuration (azure is already implemented)

### What changes are included in this PR?

System-provided GCS can be leveraged to build optional GCS features. Note that this does not support Meson to automatically download GCS - GCS is a rather complicated CMake project and does not work well in the Meson/CMake subproject bridge. Another PR will have to implement that, due to its complexity (or ideally a Meson wrap for GCS is created)

### Are these changes tested?

Yes

### Are there any user-facing changes?

No
* GitHub Issue: #46147

Authored-by: Will Ayd <william.ayd@icloud.com>
Signed-off-by: Sutou Kouhei <kou@clear-code.com>