GH-48000: [CI][Release] Publish RC GitHub Release as draft to allow immutable releases (#48059)
### Rationale for this change

GitHub allows repositories to set up immutable GitHub releases. Immutable releases won't allow for artifacts to be published to the release once the release is published and out of draft-mode. The release immutability has to be set up per repository (or organization).

### What changes are included in this PR?

Create the Release Candidate GitHub release as draft in order to allow for immutable releases to be set up on the repository.

Addition of artifacts will be available to the RC release but won't be available to the official release as we upload all artifacts as part of the release creation.

Publish the RC release once no new artifacts have to be uploaded.

### Are these changes tested?

Not as part of a release but I've created a release on my fork and tested the workflow. See:
https://github.com/raulcd/arrow/releases/tag/raul-test-release-immutable

I've validated that when immutable releases are configured on the repository:
- artifacts can be uploaded when the release is in draft mode.
- Release can be edited on draft mode
- Once the draft mode is removed no new artifacts can be uploaded
- The release can be edited (add final release notes) when the release is not in draft mode, allowing for our workflow of the release to add the final release notes without problems.

### Are there any user-facing changes?

No but users will be able to validate the artifacts have not been updated/changed since the release was done due to immutability.

* GitHub Issue: #48000

Authored-by: Raúl Cumplido <raulcumplido@gmail.com>
Signed-off-by: Raúl Cumplido <raulcumplido@gmail.com>