Make StructArray::field() thread-safe
Before this change it was possible for two threads calling `field()`
with the same index at the same time to cause a race on the stored
entry in `boxed_fields_`. I.e. if a second thread goes into the path
that calls `MakeArray` before the first thread stored its own new
array, the second thread would also write to the same shared_ptr and
invalidate the shared_ptr from the first thread, thereby also
invalidating the returned reference.