GH-47728: [Python] Check the source argument in parquet.read_table (#48008)
### Rationale for this change
See #47728. Check `source` argument in `pyarrow.parquet.read_table` if `pyarrow.dataset` is not available.

### What changes are included in this PR?
Check the `source` argument, raise `ValueError` if the `source` argument is either a list of `.parquet` files or a directory.

### Are these changes tested?
Yes

### Are there any user-facing changes?
No

In case if the `source` argument is a directory, I decided not to check it directly, but to catch the exceptions coming from the `fs.open_input_file`, since it already checks for it, and add extra exception on top of the stack that explains the actual reason.
* GitHub Issue: #47728

Authored-by: Bogdan Romenskii <rmnsk@seznam.cz>
Signed-off-by: Raúl Cumplido <raulcumplido@gmail.com>