GH-48105: [C++][Parquet][IPC] Cap allocated memory when fuzzing (#48108)
### Rationale for this change

OSS-Fuzz will trigger an out-of-memory crash if the allocated memory goes beyond a predefined limit (usually 2560 MB, though that can be configured). For Parquet and IPC, it is legitimate to allocate a lot of memory when decompressing data, though, so that can happen on both valid and invalid input files.

Unfortunately, OSS-Fuzz checks for this memory limit not by instrumenting malloc and having it return NULL when the limit is reached, but by checking allocated memory periodically from a separate thread. This can be solved by implementing our custom allocator with an upper limit, exactly how the mupdf project did in https://github.com/google/oss-fuzz/issues/1830

### What changes are included in this PR?

1. Implement a `CappedMemoryPool`
2. Use the `CappedMemoryPool` with a hardcoded limit in the Parquet and IPC fuzz targets

### Are these changes tested?

Yes, by additional unit tests.

### Are there any user-facing changes?

No.

* GitHub Issue: #48105

Authored-by: Antoine Pitrou <antoine@python.org>
Signed-off-by: Antoine Pitrou <antoine@python.org>