GH-33473 [Python] Fix KeyError on Pandas roundtrip with RangeIndex in MultiIndex (#39983)
### Rationale for this change
Fixes bug when round-tripping to Pandas with a specific combination of indices

### Are these changes tested?
Yes

### Are there any user-facing changes?
No

I don't know if this counts as a a "Critical Fix". Without it, `Table.from_pandas()` can return a table which cannot be converted back with `table.to_pandas()` due to a column missing from the `"pandas"` field in the table metadata. 
* Closes: #33473

Authored-by: Eirik B. Stavestrand <eirik.b.stavest@gmail.com>
Signed-off-by: Will Ayd <william.ayd@icloud.com>