GH-48062: [C++] Fix null pointer dereference in MakeExecBatch (#48063)
### Rationale for this change

The `arrow::compute::MakeExecBatch` function calls `DataType::id()` on `partial.type()` which could be a null pointer when `partial` is a datum representing a table. `MakeExecBatch` fails to check for this and thus null pointer dereference indeed happen. 

### What changes are included in this PR?

This patch adds a simple check in `MakeExecBatch` before calling `DataType::id()` to make sure it won't gets called on a null pointer.

### Are these changes tested?

Yes. This patch updates the unit test `ExpressionUtils.MakeExecBatch` and includes a case that calls `MakeExecBatch` with a table.

### Are there any user-facing changes?

No.

This PR indeed resolve a crash problem, but I'm not quite sure whether this should be classified as a "critical fix".

Resolve #48062 .
* GitHub Issue: #48062

Authored-by: Sirui Mu <msrlancern@gmail.com>
Signed-off-by: Antoine Pitrou <antoine@python.org>