GH-47371: [Python] Fix Numba-CUDA installation in CI
The extra that specifies the CUDA version for Numba_CUDA is required for
the CUDA Python bindings to be installed as a Numba-CUDA dependency.

This commit adds the appropriate extra to the Numba-CUDA installation,
so that the bindings will be installed in the CI environment. This
should allow the Numba-CUDA interop tests to run in CI again, instead of
getting skipped.