GH-47371, GH-48281: [Python][CI] Fix Numba-CUDA interop (#48284)
### Rationale for this change

The extra that specifies the CUDA version for Numba-CUDA is required for the CUDA Python bindings to be installed as a Numba-CUDA dependency. Presently the Numba-CUDA interop tests are skipped, and it would be preferable to enable them again and ensure the test suite doesn't skip them due to lack of dependencies.

### What changes are included in this PR?

- Addition of the appropriate extra to the Numba-CUDA installation, so that the bindings will be installed in the CI environment.
- Removal of the skip for the Numba-CUDA interop tests.

### Are these changes tested?

Yes (assuming the CI setup works correctly again with these changes)

### Are there any user-facing changes?

No.
* GitHub Issue: #47371
* GitHub Issue: #48281

Authored-by: Graham Markall <gmarkall@nvidia.com>
Signed-off-by: Antoine Pitrou <antoine@python.org>