GH-48248: [C++] Use FetchContent for bundled gRPC (#48250)
### Rationale for this change

As a follow up of requiring a minimum CMake version >= 3.25 we discussed moving our dependencies from ExternalProject to FetchContent. This can simplify our third party dependency management.

### What changes are included in this PR?

The general change is moving gRPC from `ExternalProject` to `FetchContent`. 

We can clean up previously required installation on RE2 and C-Ares. We can't do it for Abseil or protobuf as those are also used on other dependencies. We will be able to clean those later.

### Are these changes tested?

Yes, the changes are tested locally and on CI.

### Are there any user-facing changes?

No

* GitHub Issue: #48248

Authored-by: Raúl Cumplido <raulcumplido@gmail.com>
Signed-off-by: Raúl Cumplido <raulcumplido@gmail.com>