GH-48316: [C++] Use FetchContent for bundled nlohmann-json (#48320)
### Rationale for this change

As a follow up of requiring a minimum CMake version >= 3.25 we discussed moving our dependencies from ExternalProject to FetchContent. This can simplify our third party dependency management.

### What changes are included in this PR?

The general change is moving nlohmann-json from `ExternalProject` to `FetchContent`. 

The expectation is that we will be able to clean up installation once google-cloud-cpp is also migrated.

### Are these changes tested?

Yes, the changes are tested locally and on CI.

### Are there any user-facing changes?

No

Authored-by: Raúl Cumplido <raulcumplido@gmail.com>
Signed-off-by: Sutou Kouhei <kou@clear-code.com>