GH-48416: [Packaging][CI] Use custom orc_for_bundling when using FetchContent to avoid ar issues with + symbol on path (#48430)
### Rationale for this change

The centos-9 jobs have started failing to bundled ORC when using FetchContent due to subfolder including the `+` symbol on path. This fails when bundling it with `ar`.

### What changes are included in this PR?

Generate new `orc_for_bundling` library not under the `c++` directory and add that one to `ARROW_BUNDLED_STATIC_LIBS`. This is a similar approach we take for `grpc++`

### Are these changes tested?

Yes, via CI

### Are there any user-facing changes?

No, they shouldn't as this is just for adding it to the `libarrow_bundled` dependencies

* GitHub Issue: #48416

Authored-by: Raúl Cumplido <raulcumplido@gmail.com>
Signed-off-by: Sutou Kouhei <kou@clear-code.com>