GH-41246: [C++][Python] Simplify nested field encryption configuration (#45462)
### Rationale for this change

Columns can b encrypted with individual keys. For this, the column name have to be set in `EncryptionConfiguration::column_keys`. This poses the following challenges for columns with nested fields like `MapType`, `ListType`, and `StructType`. Encrypting a column of such type requires providing an encryption key for all nested (leaf) fields. Ideally, the column name should be sufficient (as it is for any other data type) to encrypt all nested fields.

### What changes are included in this PR?

The column name can be used to encrypt all nested fields of `StrutType`, `MapType`, and `ListType` columns with the same encryption key. The current column naming scheme can still be used for backward compatibility.

### Are these changes tested?
Tested in C++ and Python.

### Are there any user-facing changes?
Column encryption can be configured with less code and more intuitive naming.

Documentation and examples updated.

Fixes #41246.
* GitHub Issue: #41246

Authored-by: Enrico Minack <github@enrico.minack.dev>
Signed-off-by: Antoine Pitrou <antoine@python.org>