GH-41364: [GLib][Ruby] Allow passing thread pool to ExecutePlan (#48462)
### Rationale for this change

Aggregators like `first` and `last` are unusable in Ruby because they don't work when the execution plan is executed using multiple threads.

### What changes are included in this PR?

This adds the `ThreadPool` class to be able create a thread pool with a single thread. This can then be passed when creating an `ExecuteContext`, which in turn can be passed when creating an `ExecutePlan`.

### Are these changes tested?

A Ruby test that shows that the `first` aggregator works.

### Are there any user-facing changes?

A new `GArrowThreadPool` class, and changed signatures of the functions `garrow_execute_context_new` and `garrow_execute_plan_new`. However since the new arguments are nullable, it is backwards compatible for the Ruby API.

**This PR includes breaking changes to public APIs.**

* GitHub Issue: #41364

Authored-by: Sten Larsson <sten@burtcorp.com>
Signed-off-by: Sutou Kouhei <kou@clear-code.com>